+++ packages/web/src/App.tsx
@@ -11,7 +11,6 @@ import { ChannelsPage } from "./pages/ChannelsPage";
 import { SubagentsPage } from "./pages/SubagentsPage";
 import { AgentsPage } from "./pages/AgentsPage";
 import { ApiPage } from "./pages/ApiPage";
-import { ProjectsPage } from "./pages/ProjectsPage";
 import { ProjectDetailPage } from "./pages/ProjectDetailPage";
 import { LoginPage } from "./pages/LoginPage";
 import { AuthProvider, useAuth } from "./auth";
@@ -48,9 +47,7 @@ function AppRoutes() {
           <Route path="/" element={<Navigate to="/chat" replace />} />
           <Route path="/chat" element={<ChatPage />} />
           <Route path="/chat/:sessionId" element={<ChatPage />} />
-          <Route path="/projects" element={<ProjectsPage />} />
           <Route path="/projects/:id" element={<ProjectDetailPage />} />
-          <Route path="/projects/:id/edit" element={<ProjectsPage />} />
           <Route path="/tasks" element={<TasksPage />} />
           <Route path="/channels" element={<ChannelsPage />} />
           <Route path="/subagents" element={<SubagentsPage />} />